+++ forward_engineering/generateViewScript.js
@@ -1,0 +1,9 @@
+const { generateScript } = require('./generateScript');
+
+const generateViewScript = (...args) => {
+	return generateScript(...args);
+};
+
+module.exports = {
+	generateViewScript,
+};
